UTArray: fix last warning.

+++ ext/uthash/utarray.h
@@ -219,7 +219,7 @@ typedef struct {
 #define utarray_next(a,e) (((e)==NULL) ? utarray_front(a) : ((((a)->i) > (utarray_eltidx(a,e)+1)) ? _utarray_eltptr(a,(size_t)utarray_eltidx(a,e)+1) : NULL))
 #define utarray_prev(a,e) (((e)==NULL) ? utarray_back(a) : ((utarray_eltidx(a,e) > 0) ? _utarray_eltptr(a,(size_t)utarray_eltidx(a,e)-1) : NULL))
 #define utarray_back(a) (((a)->i) ? (_utarray_eltptr(a,(a)->i-1)) : NULL)
-#define utarray_eltidx(a,e) (((char*)(e) >= (char*)((a)->d)) ? (ssize_t)(((char*)(e) - (char*)((a)->d))/(a)->icd.sz) : (ssize_t)-1)
+#define utarray_eltidx(a,e) (((char*)(e) >= (char*)((a)->d)) ? (ssize_t)((size_t)((char*)(e) - (char*)((a)->d))/(a)->icd.sz) : (ssize_t)-1)
 
 /* last we pre-define a few icd for common utarrays of ints and strings */
 static void utarray_str_cpy(void *dst, const void *src) {
